netlify/functions/requirements.txt
flask
flask-cors
newspaper3k
textblob
nltk
pickle
flask-lambda